add tests (#3)
Adds test to check the contents of various spans, and an e2e test for a libhoney event

In doing so:

I've removed the status.message field since that was returning just an empty string.
I've fixed the all 0s trace-id. We now get trace-ids, but they're not passed through the from the client request to the server-side test. So, I'll keep brainstorming for how to test that part of things.