export span Attributes (#20)
This translates the attributes to fields on the span.

In developing this I noticed 2 issues that I'll file.

The SpanData notes that "The values of Attributes each have type string, bool, or int64."
Which isn't true. They can also be the core.Value type.
The Jaeger exporter won't be able to export Attributes of that core.Value type. I'll point to this PR in those issues.