Add method for reading from responses queue. (#70)
The `onError` hook function is called when the exporter runs into an
error sending data to Honeycomb. Many errors however, such as common
transport errors, are logged asynchronously in the response queue.
Adding a method called `RunErrorLogger` to read from the response
queue and call the `onError` hook when errors are encountered. The
function accepts a `context.Context` which can be used to cancel or
timeout the operation.