Transcribe attributes from span links to events (#96)
Links attached to OpenTelemetry spans carry attributes like spans and events. Since we publish links as Honeycomb events, and events all carry fields, transcribe attributes on links as fields on the corresponding Honeycomb events.

Include OpenTelemetry resource attributes as the underlay, and shadow those resource attributes if necessary with the link-level attributes, as we did for spans and span events in #71.